+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Innmind\Framework\Http\Route\Reference`
 - `Innmind\Framework\Application::mapRoute()`
 - `Innmind\Framework\Application::routes(class-string<Innmind\Framework\Http\Route\Reference>)`
+- `Innmind\Framework\Application::recoverRouteError()`
 
 ### Changed